requirements.txt
gurobipy==10.0.3
pulp==2.7.0
geopandas==0.14.2
shapely==2.0.3
numpy==1.26.3
pandas==2.1.4
scipy==1.12.0
matplotlib==3.8.2
seaborn==0.13.2
folium==0.15.1
libpysal==4.9.2
esda==2.6.0
splot==1.1.5
contextily==1.5.0
pyproj==3.6.1
scikit-learn==1.4.0
tqdm==4.66.1
pyyaml==6.0.1
pytest==7.4.4
jupyter==1.0.0
ipykernel==6.29.0
